fixed refactoring and linting
